fix token bug post last update that removed custom claims
